@@ -9,7 +9,7 @@ import { Component, OnInit } from '@angular/core';
 export class MenuComponent implements OnInit {
   //isUserLoggedIn: boolean = false;
 
-  constructor(private hardcodedAuthenticationService 
+  constructor(public hardcodedAuthenticationService
     : HardcodedAuthenticationService) { }
 
   ngOnInit() {
