@@ -9,32 +9,32 @@ import { Todo } from '../../list-todos/list-todos.component';
 export class TodoDataService {
 
   constructor(
-    private http:HttpClient
+    private http: HttpClient
   ) { }
 
-  retrieveAllTodos(username) {
+  retrieveAllTodos(username: string) {
     return this.http.get<Todo[]>(`${TODO_JPA_API_URL}/users/${username}/todos`);
     //console.log("Execute Hello World Bean Service")
   }
 
-  deleteTodo(username, id){
+  deleteTodo(username: string, id: number) {
     return this.http.delete(`${TODO_JPA_API_URL}/users/${username}/todos/${id}`);
   }
 
-  retrieveTodo(username, id){
+  retrieveTodo(username: string, id: number) {
     return this.http.get<Todo>(`${TODO_JPA_API_URL}/users/${username}/todos/${id}`);
   }
 
-  updateTodo(username, id, todo){
+  updateTodo(username: string, id: number, todo: Todo) {
     return this.http.put(
-          `${TODO_JPA_API_URL}/users/${username}/todos/${id}`
-                , todo);
+      `${TODO_JPA_API_URL}/users/${username}/todos/${id}`
+      , todo);
   }
 
-  createTodo(username, todo){
+  createTodo(username: string, todo: Todo) {
     return this.http.post(
-              `${TODO_JPA_API_URL}/users/${username}/todos`
-                , todo);
+      `${TODO_JPA_API_URL}/users/${username}/todos`
+      , todo);
   }
 
 }
