@@ -10,8 +10,8 @@ import { Todo } from '../list-todos/list-todos.component';
 })
 export class TodoComponent implements OnInit {
 
-  id:number
-  todo: Todo
+  id: number = 0;
+  todo: Todo = new Todo(this.id, '', false, new Date());
 
   constructor(
     private todoService: TodoDataService,
@@ -20,36 +20,36 @@ export class TodoComponent implements OnInit {
   ) { }
 
   ngOnInit() {
-    
+
     this.id = this.route.snapshot.params['id'];
-    
-    this.todo = new Todo(this.id,'',false,new Date());
-    
-    if(this.id!=-1) {
+
+    this.todo = new Todo(this.id, '', false, new Date());
+
+    if (this.id != -1) {
       this.todoService.retrieveTodo('in28minutes', this.id)
-          .subscribe (
-            data => this.todo = data
-          )
+        .subscribe(
+          data => this.todo = data
+        )
     }
   }
 
   saveTodo() {
-    if(this.id == -1) { //=== ==
+    if (this.id == -1) { //=== ==
       this.todoService.createTodo('in28minutes', this.todo)
-          .subscribe (
-            data => {
-              console.log(data)
-              this.router.navigate(['todos'])
-            }
-          )
+        .subscribe(
+          data => {
+            console.log(data)
+            this.router.navigate(['todos'])
+          }
+        )
     } else {
       this.todoService.updateTodo('in28minutes', this.id, this.todo)
-          .subscribe (
-            data => {
-              console.log(data)
-              this.router.navigate(['todos'])
-            }
-          )
+        .subscribe(
+          data => {
+            console.log(data)
+            this.router.navigate(['todos'])
+          }
+        )
     }
   }
 
