@@ -18,36 +18,36 @@ import { Component, OnInit } from '@angular/core';//'./app.component';
 export class WelcomeComponent implements OnInit {
 
   message = 'Some Welcome Message'
-  welcomeMessageFromService:string
+  welcomeMessageFromService: string = ''
   name = ''
   //String message = "Some Welcome Message"
-  
+
   //public SpringBootFirstWebApplication() {	
 
   //ActivatedRoute
   constructor(
-    private route:ActivatedRoute,
-    private service:WelcomeDataService) { 
+    private route: ActivatedRoute,
+    private service: WelcomeDataService) {
 
   }
 
   // void init() {
-  ngOnInit(){
+  ngOnInit() {
     //COMPILATION ERROR this.message = 5
     //console.log(this.message)
     // console.log(this.route.snapshot.params['name'])
     this.name = this.route.snapshot.params['name'];
-    
+
   }
 
   getWelcomeMessage() {
     //console.log(this.service.executeHelloWorldBeanService());
-    
+
     this.service.executeHelloWorldBeanService().subscribe(
       response => this.handleSuccessfulResponse(response),
       error => this.handleErrorResponse(error)
     );
-    
+
     //console.log('last line of getWelcomeMessage')
 
     //console.log("get welcome message");
@@ -55,25 +55,25 @@ export class WelcomeComponent implements OnInit {
 
   getWelcomeMessageWithParameter() {
     //console.log(this.service.executeHelloWorldBeanService());
-    
+
     this.service.executeHelloWorldServiceWithPathVariable(this.name).subscribe(
       response => this.handleSuccessfulResponse(response),
       error => this.handleErrorResponse(error)
     );
-    
+
     //console.log('last line of getWelcomeMessage')
 
     //console.log("get welcome message");
   }
 
 
-  handleSuccessfulResponse(response){
+  handleSuccessfulResponse(response: any) {
     this.welcomeMessageFromService = response.message
     //console.log(response);
     //console.log(response.message);
   }
 
-  handleErrorResponse(error) {
+  handleErrorResponse(error: any) {
     //console.log(error);
     //console.log(error.error);
     //console.log(error.error.message);
@@ -86,5 +86,5 @@ export class Class1 {
 }
 
 export class Class2 {
-  
+
 }
