@@ -13,16 +13,16 @@ public class RestfulWebServicesApplication {
 		SpringApplication.run(RestfulWebServicesApplication.class, args);
 	}
 
-	//http://localhost:3000/ to 8080
+	//http://localhost:4200/ to 8080
 	//Cross Origin Requests
-	//Allow all requests only from http://localhost:3000/
+	//Allow all requests only from http://localhost:4200/
 	@Bean
 	public WebMvcConfigurer corsConfigurer() {
 		return new WebMvcConfigurer() {
 			public void addCorsMappings(CorsRegistry registry) {
 				registry.addMapping("/**")
 						.allowedMethods("*")
-						.allowedOrigins("http://localhost:3000");
+						.allowedOrigins("http://localhost:4200");
 			}
 		};
 	}
